modified

+++ src/entity/User.entity.ts
@@ -13,7 +13,7 @@ export class User extends BaseModel {
   email: string;
   @Column({ unique: true, nullable: true })
   account_number: string;
-  @Column({ type: "numeric", default: 0 })
+  @Column({ type: "bigint", default: 0 })
   balance: number;
   @Column()
   first_name: string;
